// packages/openapi-ts-tests/tanstack-query/v6/__snapshots__/fetch/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseUrl: `${string}://${string}` | (string & {});
};

export type Pet = {
  id: number;
  name: string;
};

export type ApiError = {
  message: string;
};

export type ListPetsData = {
  body?: never;
  path?: never;
  query?: {
    page?: number;
    limit?: number;
  };
  url: '/pets';
};

export type ListPetsResponses = {
  /**
   * A page of pets
   */
  200: Array<Pet>;
};

export type ListPetsResponse = ListPetsResponses[keyof ListPetsResponses];

export type AddPetData = {
  body: Pet;
  path?: never;
  query?: never;
  url: '/pets';
};

export type AddPetErrors = {
  /**
   * Request failed
   */
  400: ApiError;
};

export type AddPetError = AddPetErrors[keyof AddPetErrors];

export type AddPetResponses = {
  /**
   * Created pet
   */
  200: Pet;
};

export type AddPetResponse = AddPetResponses[keyof AddPetResponses];

export type GetPetData = {
  body?: never;
  path: {
    petId: number;
  };
  query?: never;
  url: '/pets/{petId}';
};

export type GetPetErrors = {
  /**
   * Request failed
   */
  404: ApiError;
};

export type GetPetError = GetPetErrors[keyof GetPetErrors];

export type GetPetResponses = {
  /**
   * A pet
   */
  200: Pet;
};

export type GetPetResponse = GetPetResponses[keyof GetPetResponses];